Откатывает мастер к прошлому заданию
